@@ -1,11 +1,11 @@
 name: device_info_csx
 description: Flutter plugin providing detailed information about the device
   (make, model, etc.), and Android or iOS version the app is running on.
-version: 0.0.1
+version: 0.0.2
 homepage: https://github.com/KirstenDunst/FlutterPlugins/tree/main/device_info_csx
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
