@@ -24,31 +24,31 @@ class IosDeviceInfo {
   });
 
   /// Device name.
-  final String name;
+  final String? name;
 
   /// The name of the current operating system.
-  final String systemName;
+  final String? systemName;
 
   /// The current operating system version.
-  final String systemVersion;
+  final String? systemVersion;
 
   /// Device model.
-  final String model;
+  final String? model;
 
   /// Localized name of the device model.
-  final String localizedModel;
+  final String? localizedModel;
 
   /// Unique UUID value identifying the current device.
-  final String identifierForVendor;
+  final String? identifierForVendor;
 
   /// `false` if the application is running in a simulator, `true` otherwise.
-  final bool isPhysicalDevice;
+  final bool? isPhysicalDevice;
 
   /// Operating system information derived from `sys/utsname.h`.
-  final IosUtsname utsname;
+  final IosUtsname? utsname;
 
   /// device storage detail.
-  final Storage storage;
+  final Storage? storage;
 
   /// Deserializes from the map message received from [_kChannel].
   static IosDeviceInfo fromMap(Map<String, dynamic> map) {
@@ -74,8 +74,8 @@ class IosDeviceInfo {
     data['localizedModel'] = localizedModel;
     data['identifierForVendor'] = identifierForVendor;
     data['isPhysicalDevice'] = isPhysicalDevice;
-    data['utsname'] = utsname.toJson();
-    data['storage'] = storage.toJson();
+    data['utsname'] = utsname!.toJson();
+    data['storage'] = storage!.toJson();
     return data;
   }
 }
@@ -92,19 +92,19 @@ class IosUtsname {
   });
 
   /// Operating system name.
-  final String sysname;
+  final String? sysname;
 
   /// Network node name.
-  final String nodename;
+  final String? nodename;
 
   /// Release level.
-  final String release;
+  final String? release;
 
   /// Version level.
-  final String version;
+  final String? version;
 
   /// Hardware type (e.g. 'iPhone7,1' for iPhone 6 Plus).
-  final String machine;
+  final String? machine;
 
   /// Deserializes from the map message received from [_kChannel].
   static IosUtsname _fromMap(Map<String, dynamic> map) {
@@ -130,13 +130,13 @@ class IosUtsname {
 
 class Storage {
   // 已使用的ram，单位byte
-  num ramUseB;
+  num? ramUseB;
   // 全部的ram容量，单位byte
-  num ramAllB;
+  num? ramAllB;
   // 已使用的rom，单位byte
-  num romUseB;
+  num? romUseB;
   // 全部的rom容量，单位byte
-  num romAllB;
+  num? romAllB;
 
   Storage({
     this.ramUseB,
