@@ -1,10 +1,16 @@
 package cn.brainco.device_info_csx;
 
 import android.annotation.SuppressLint;
+import android.app.ActivityManager;
 import android.content.ContentResolver;
+import android.content.Context;
 import android.content.pm.FeatureInfo;
 import android.content.pm.PackageManager;
 import android.os.Build;
+import android.os.Debug;
+import android.os.Environment;
+import android.os.Process;
+import android.os.StatFs;
 import android.provider.Settings;
 
 import androidx.annotation.NonNull;
@@ -36,12 +42,14 @@ public class DeviceInfoCsxPlugin implements FlutterPlugin, MethodCallHandler {
 
     private ContentResolver contentResolver;
     private PackageManager packageManager;
+    private Context applicationContext;
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
         channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "device_info_csx");
         this.contentResolver = flutterPluginBinding.getApplicationContext().getContentResolver();
         this.packageManager = flutterPluginBinding.getApplicationContext().getPackageManager();
+        this.applicationContext = flutterPluginBinding.getApplicationContext();
         channel.setMethodCallHandler(this);
     }
 
@@ -90,44 +98,92 @@ public class DeviceInfoCsxPlugin implements FlutterPlugin, MethodCallHandler {
             build.put("version", version);
 
             Map<String, Object> storage = new HashMap<>();
-            storage.put("ramUseB", getRamUseB());
-            storage.put("ramAllB", getRamAllB());
-            storage.put("romUseB", getRomUseB());
-            storage.put("romAllB", getRomAllB());
-            build.put("storage", storage);
 
+            //设备RAM信息
+            inflateDeviceRAM(storage);
+
+            //设备ROM信息
+            inflateDeviceROM(storage);
+
+            //当前进程的总内存占用情况
+            inflatePss(storage);
+
+            //jvm内存使用情况
+            inflateJvmMemory(storage);
+
+            build.put("storage", storage);
             result.success(build);
         } else {
             result.notImplemented();
         }
     }
 
-    /**
-     * 已使用的ram，单位byte
-     */
-    private double getRamUseB() {
-        return 0.0;
+    private void inflateJvmMemory(Map<String, Object> storage) {
+        long maxMem = Runtime.getRuntime().maxMemory();//当前虚拟机可用的最大内存
+        long totalMem = Runtime.getRuntime().totalMemory();//当前虚拟机已分配的内存
+        long freeMemory = Runtime.getRuntime().freeMemory();//当前虚拟机已分配内存中未使用的部分
+        long currentMemUsage = totalMem - freeMemory;
+        storage.put("jvmUseB", currentMemUsage);
+        storage.put("jvmMaxB", maxMem);
+        storage.put("jvmTotalB", totalMem);
+        storage.put("jvmFreeB", freeMemory);
     }
 
-    /**
-     * 全部的ram容量，单位byte
-     */
-    private double getRamAllB() {
-        return 0.0;
+    private void inflateDeviceROM(Map<String, Object> storage) {
+        StatFs statFs = new StatFs(Environment.getDataDirectory().getPath());
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR2) {
+            long totalCounts = statFs.getBlockCountLong();
+            long availableCounts = statFs.getAvailableBlocksLong();
+            long size = statFs.getBlockSizeLong();
+            long availROMSize = availableCounts * size;
+            long totalROMSize = totalCounts * size;
+            storage.put("romAllB", totalROMSize);
+            storage.put("romUseB", totalROMSize - availROMSize);
+        } else {
+            long totalCounts = statFs.getBlockCount();
+            long availableCounts = statFs.getAvailableBlocks();
+            long size = statFs.getBlockSize();
+            long availROMSize = availableCounts * size;
+            long totalROMSize = totalCounts * size;
+            storage.put("romAllB", totalROMSize);
+            storage.put("romUseB", totalROMSize - availROMSize);
+        }
     }
 
-    /**
-     * 已使用的rom，单位byte
-     */
-    private double getRomUseB() {
-        return 0.0;
+    private void inflateDeviceRAM(Map<String, Object> storage) {
+        ActivityManager activityManager = (ActivityManager) applicationContext.getSystemService(Context.ACTIVITY_SERVICE);
+        ActivityManager.MemoryInfo memoryInfo = new ActivityManager.MemoryInfo();
+        activityManager.getMemoryInfo(memoryInfo);
+        storage.put("ramUseB", memoryInfo.totalMem - memoryInfo.availMem);
+        storage.put("ramAllB", memoryInfo.totalMem);
+        storage.put("ramAvailableB", memoryInfo.availMem);
+        storage.put("ramThreshold", memoryInfo.threshold); // availMem 的阈值，低于此值系统会开始杀进程，后台服务
+        storage.put("lowMemory", memoryInfo.lowMemory);// 是否为低内存
     }
 
-    /**
-     * 全部的rom容量，单位byte
-     */
-    private double getRomAllB() {
-        return 0.0;
+    private void inflatePss(Map<String, Object> storage) {
+        ActivityManager activityManager = (ActivityManager) applicationContext.getSystemService(Context.ACTIVITY_SERVICE);
+        Debug.MemoryInfo[] memInfo = activityManager.getProcessMemoryInfo(new int[]{Process.myPid()});
+        if (memInfo.length > 0) {
+            //AndroidQ 版本对这个 API 增加了限制，当采样率较高时，会一直返回一个相同的值
+            //包括了虚拟机的内存占用情况，还包括原生层和其它内存占用，这也是在手机连接至电脑时，使用 ADB 指令查看到的内存占用信息。
+            // TotalPss = dalvikPss + nativePss + otherPss, in KB
+            Debug.MemoryInfo info = memInfo[0];
+            long totalPss = info.getTotalPss();
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+                Map<String, String> memoryStats = info.getMemoryStats();
+                int other = Integer.parseInt(memoryStats.get("summary.private-other"));
+                storage.put("summaryJavaHeap", Integer.parseInt(memoryStats.get("summary.java-heap")) << 10);
+                storage.put("summaryNativeHeap", Integer.parseInt(memoryStats.get("summary.native-heap")) << 10);
+                storage.put("summaryCode", Integer.parseInt(memoryStats.get("summary.code")) << 10);
+                storage.put("summaryStack", Integer.parseInt(memoryStats.get("summary.stack")) << 10);
+                storage.put("summaryGraphics", Integer.parseInt(memoryStats.get("summary.graphics")) << 10);
+                storage.put("summaryPrivateOther", Integer.parseInt(memoryStats.get("summary.private-other")) << 10);
+                storage.put("summarySystem", Integer.parseInt(memoryStats.get("summary.system")) << 10);
+                storage.put("summaryTotalSwap", Integer.parseInt(memoryStats.get("summary.total-swap")) << 10);
+            }
+            storage.put("totalPssB", totalPss << 10);
+        }
     }
 
     private String[] getSystemFeatures() {
